Convert color value to float numbers instead of integers since float numbers are used

+++ src/Canvas/CanvasContext.h
@@ -118,7 +118,7 @@ namespace mural
         float getGlobalAlpha();
     };
 
-    void stringToColorRGBA(const std::string& color, int& r, int& g, int& b, float& a);
+    void stringToColorRGBA(const std::string& color, float& r, float& g, float& b, float& a);
 }
 
 #endif /* defined(__CinderCanvas__CanvasContext__) */
